Move editing profile endpoint to /users/
